requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap4==3.0.0
django-braces==1.15.0
django-crispy-forms==1.11.2
django-stdimage==6.0.1
gunicorn==20.1.0
model-mommy==2.0.0
mysql==0.0.3
mysqlclient==2.1.1
pep517==0.13.0
phonenumbers==8.13.3
Pillow==9.4.0
psycopg2-binary==2.9.1
PyMySQL==1.0.2
pytz==2022.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
whitenoise==6.2.0
